+++ medical-wiki/src/theme/MDXComponents.js
@@ -4,6 +4,7 @@ import ImageUploader from '@site/src/components/ImageUploader';
 import { ImageGallery } from '@site/src/components/ImageUploader';
 import DiagnosticBooster from '@site/src/components/DiagnosticBooster';
 import TreatmentBooster from '@site/src/components/TreatmentBooster';
+import EchoBooster from '@site/src/components/EchoBooster';
 import MermaidChart from '@site/src/components/MermaidChart';
 import DupixentCalculator from '@site/src/components/ClinicalCalc/DupixentCalculator';
 
@@ -13,6 +14,7 @@ export default {
   ImageGallery,
   DiagnosticBooster,
   TreatmentBooster,
+  EchoBooster,
   MermaidChart,
   DupixentCalculator,
 };